Move merge logic in to vcfg/merge.go
